Fix pin offsets from plug body
From mrgo0se:

The current code to calculate the pin array offsets from the plug body
appears to be incorrect. This gives slightly different distances between
the first and last pin in each row and the closest corner on the plug
body. i.e. the pin arrays are not centred on their respective sides of
the plug resulting in an asymmetric model.